tasks/process_new_tweets/requirements.txt
tweepy==4.8.0
python-dotenv==0.19.2
